Update models.py
delete method changed

+++ deals/models.py
@@ -199,13 +199,13 @@ class Store(models.Model):
         return self.subscriptions.all()
     
     def delete(self, *args, **kwargs):
-        old_image_path = os.path.join(settings.BASE_DIR, self.image_url.lstrip('/'))
-        if os.path.exists(old_image_path):
-            os.remove(old_image_path)
-        else:
-            raise FileNotFoundError(f"File not found: {old_image_path}")
-        
-        # Call the original delete method to remove the database record
+        # If the field is a Django ImageField or FileField
+        if self.image:
+            # Get the full filesystem path
+            old_image_path = self.image.path
+            if os.path.exists(old_image_path):
+                os.remove(old_image_path)
+        # Then delete the object itself
         super().delete(*args, **kwargs)
 
     def to_dict(self):
